+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - changed: (BSC) Lower minimum gas price to 0.05 gwei
+- changed: Report Piratechain engine syncRatio as download progress since birthday height
 - fixed: Zano `parseUri` currencyCode validation
 
 ## 4.63.0 (2025-11-04)